+++ src/StrongTypes.Api.IntegrationTests/Tests/BindingTests/BindingTestAsserts.cs
@@ -1,0 +1,33 @@
+using System.Net;
+using System.Net.Http.Json;
+using System.Text.Json;
+using Xunit;
+
+namespace StrongTypes.Api.IntegrationTests.Tests;
+
+internal static class BindingTestAsserts
+{
+    /// <summary>
+    /// Asserts the response is a 400 with a <c>ValidationProblemDetails</c>
+    /// payload whose <c>errors</c> map contains an entry for
+    /// <paramref name="expectedField"/> (case-insensitive — header names use
+    /// kebab-case while query/form fields use camelCase).
+    /// </summary>
+    internal static async Task AssertValidationProblem(HttpResponseMessage response, string expectedField)
+    {
+        Assert.Equal(HttpStatusCode.BadRequest, response.StatusCode);
+        var problem = await response.Content.ReadFromJsonAsync<JsonElement>(TestContext.Current.CancellationToken);
+        var errors = problem.GetProperty("errors");
+        Assert.Equal(JsonValueKind.Object, errors.ValueKind);
+        var found = false;
+        foreach (var prop in errors.EnumerateObject())
+        {
+            if (string.Equals(prop.Name, expectedField, StringComparison.OrdinalIgnoreCase))
+            {
+                found = true;
+                break;
+            }
+        }
+        Assert.True(found, $"Expected ValidationProblemDetails to include error for '{expectedField}'. Actual: {problem}");
+    }
+}
